httpstoppable: Remove obsoleted deprecated package.
Go 1.8 added native support for stopping a server in net/http, and
should be used instead.
