+++ README.md
@@ -24,7 +24,6 @@ Directories
 | [gopathutil](https://godoc.org/github.com/shurcooL/go/gopathutil)                                     | Package gopathutil provides tools to operate on GOPATH workspace.                                                                                                 |
 | [gopherjs_http](https://godoc.org/github.com/shurcooL/go/gopherjs_http)                               | Package gopherjs_http provides helpers for compiling Go using GopherJS and serving it over HTTP.                                                                  |
 | [gopherjs_http/jsutil](https://godoc.org/github.com/shurcooL/go/gopherjs_http/jsutil)                 | Package jsutil provides utility functions for interacting with native JavaScript APIs.                                                                            |
-| [httpstoppable](https://godoc.org/github.com/shurcooL/go/httpstoppable)                               | Package httpstoppable provides ListenAndServe like http.ListenAndServe, but with ability to stop it.                                                              |
 | [importgraphutil](https://godoc.org/github.com/shurcooL/go/importgraphutil)                           | Package importgraphutil augments "golang.org/x/tools/refactor/importgraph" with a way to build graphs ignoring tests.                                             |
 | [indentwriter](https://godoc.org/github.com/shurcooL/go/indentwriter)                                 | Package indentwriter implements an io.Writer wrapper that indents every non-empty line with specified number of tabs.                                             |
 | [ioutil](https://godoc.org/github.com/shurcooL/go/ioutil)                                             | Package ioutil provides a WriteFile func with an io.Reader as input.                                                                                              |
